requirements.txt
requests
pandas
pyarrow
sqlalchemy
python-dotenv
psycopg2-binary
prefect
